function/data-upserter/pyproject.toml
[tool.poetry]
name = "data-upserter"
version = "1.0"
description = "Upserts data from a staging table to a main table"
authors = ["Johnson Luong <joluong@firstam.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
